Bumped version to 2.1.1 Renamed file useBarDrag to createBarDrag Re-enabled default import Added plugin for auto-injecting CSS in vite build Updated README
