+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@infectoone/vue-ganttastic",
-  "version": "3.0.0",
+  "version": "2.1.1",
   "description": "A simple and customizable Gantt chart component for Vue.js",
   "author": "Marko Zunic (@infectoone)",
   "scripts": {
@@ -23,6 +23,7 @@
   ],
   "devDependencies": {
     "@rushstack/eslint-patch": "^1.1.4",
+    "@senojs/rollup-plugin-style-inject": "^0.1.1",
     "@types/node": "^16.11.58",
     "@types/postcss-preset-env": "^7.7.0",
     "@vitejs/plugin-vue": "^3.1.0",
