+++ src/composables/createBarDrag.ts
