+++ src/main.ts
@@ -1,5 +1,5 @@
 import { createApp } from "vue"
 import Playground from "./GanttPlayground.vue"
-import { ganttastic } from "./index"
+import ganttastic from "./index"
 
 createApp(Playground).use(ganttastic).mount("#app")
